Refactored charts folder
