+++ web/src/features/channel/eventsChart/EventsChart.tsx
@@ -1,9 +1,9 @@
 // https://www.pluralsight.com/guides/using-d3.js-inside-a-react-app
-import { useD3 } from "../charts/useD3";
+import { useD3 } from "../../charts/useD3";
 import React, { useEffect } from "react";
 import { Selection } from "d3";
-import { ChartCanvas, AreaPlot, EventsPlot, LinePlot } from "../charts/charts";
-import "../charts/chart.scss";
+import { ChartCanvas, AreaPlot, EventsPlot, LinePlot } from "../../charts/charts";
+import "../../charts/chart.scss";
 
 type EventsChart = {
   data: any[];
