+++ web/src/features/charts/chart.scss
