+++ web/src/features/charts/plots/abstract.ts
