+++ web/src/features/charts/plots/bar.ts
