+++ web/src/features/charts/plots/eventIcons.ts
